Revert "Feature/ed 1457 users"
