+++ src/App.tsx
@@ -1,7 +1,10 @@
 import type { GetProp, MenuProps } from 'antd';
 
 import { UnorderedListOutlined, UserOutlined } from '@ant-design/icons';
-import { useEffect } from 'react';
+import { Layout, Menu } from 'antd';
+import { Content } from 'antd/es/layout/layout';
+import Sider from 'antd/es/layout/Sider';
+import { useEffect, useState } from 'react';
 import {
   createBrowserRouter,
   Link,
@@ -11,53 +14,57 @@ import {
   useNavigate
 } from 'react-router';
 
-import { MainLayout } from '@/components/MainLayout.tsx';
 import { utilsTokens } from '@/helpers/tokenService.ts';
 import { ProfilePage } from '@/pages/Profile/ProfilePage.tsx';
 import { SignInPage } from '@/pages/SignIn/SignInPage.tsx';
 import { SignUpPage } from '@/pages/SignUp/SignUpPage.tsx';
 import { TodosPage } from '@/pages/Todos/TodosPage.tsx';
-import { UserPage } from '@/pages/User/UserPage.tsx';
-import { UsersPage } from '@/pages/Users/UsersPage.tsx';
-import { selectAuthIsAuth, selectAuthProfile } from '@/store/selectors.ts';
-import { authActions, getProfile, refreshAccessToken } from '@/store/slices/authSlice.ts';
+import { selectAuthIsAuth } from '@/store/selectors.ts';
+import { authActions, refreshAccessToken } from '@/store/slices/authSlice.ts';
 import { useAppDispatch, useAppSelector } from '@/store/store.ts';
 
 import './App.css';
 
 type MenuItem = GetProp<MenuProps, 'items'>[number];
+const items: MenuItem[] = [
+  {
+    key: '/todos',
+    icon: <UnorderedListOutlined />,
+    label: <Link to='/todos'>Todo</Link>
+  },
+  {
+    key: '/profile',
+    icon: <UserOutlined />,
+    label: <Link to='/profile'>Profile</Link>
+  }
+];
 
 export const App = () => {
-  const dispatch = useAppDispatch();
+  const [isOpen, setIsOpen] = useState<boolean>(false);
   const location = useLocation();
-  const { data } = useAppSelector(selectAuthProfile);
-  const isAuth = useAppSelector(selectAuthIsAuth);
-
-  useEffect(() => {
-    if (isAuth) {
-      dispatch(getProfile());
-    }
-  }, [isAuth, location.pathname]);
-  const items: MenuItem[] = [
-    {
-      key: '/todos',
-      icon: <UnorderedListOutlined />,
-      label: <Link to='/todos'>Todo</Link>
-    },
-    {
-      key: '/profile',
-      icon: <UserOutlined />,
-      label: <Link to='/profile'>Profile</Link>
-    }
-  ];
-  if (data?.roles.some((role) => role === 'ADMIN' || role === 'MODERATOR')) {
-    items.push({ key: '/users', icon: <UserOutlined />, label: <Link to='/users'>Users</Link> });
-  }
-
   return (
-    <MainLayout items={items}>
-      <Outlet />
-    </MainLayout>
+    <>
+      <Layout style={{ minHeight: '100vh' }}>
+        <Sider
+          breakpoint={'md'}
+          collapsed={isOpen}
+          collapsible
+          onCollapse={(value) => setIsOpen(value)}
+        >
+          <Menu
+            defaultSelectedKeys={[location.pathname]}
+            items={items}
+            theme='dark'
+            mode='inline'
+          />
+        </Sider>
+        <Layout>
+          <Content>
+            <Outlet />
+          </Content>
+        </Layout>
+      </Layout>
+    </>
   );
 };
 
@@ -92,23 +99,6 @@ const ProtectedRoute = () => {
   }
   return <Outlet />;
 };
-// const ProtectedAdminOrModeratorRoute = () => {
-//   const { data } = useAppSelector(selectAuthProfile);
-//   const navigate = useNavigate();
-//
-//   useEffect(() => {
-//     if (!data?.roles.some((role) => role === 'ADMIN' || role === 'MODERATOR')) {
-//       navigate('/todos');
-//       console.log('yep');
-//     }
-//   }, [data]);
-//
-//   const isAdminOrModer = data?.roles.some((role) => role === 'ADMIN' || role === 'MODERATOR');
-//   if (!isAdminOrModer) {
-//     return null;
-//   }
-//   return <Outlet />;
-// };
 
 // const protectedLoader = async () => {
 //   console.log('render');
@@ -150,18 +140,6 @@ export const router = createBrowserRouter([
           {
             path: '/profile',
             Component: ProfilePage
-          },
-          {
-            children: [
-              {
-                path: '/users',
-                Component: UsersPage
-              },
-              {
-                path: '/users/:userId',
-                Component: UserPage
-              }
-            ]
           }
         ]
       }
