+++ src/pages/Profile/ProfilePage.tsx
@@ -1,9 +1,10 @@
 import { Button, Flex, Typography } from 'antd';
+import { useEffect } from 'react';
 import { useNavigate } from 'react-router';
 
 import { utilsTokens } from '@/helpers/tokenService.ts';
-import { selectAuthProfile } from '@/store/selectors.ts';
-import { authActions, logoutUser } from '@/store/slices/authSlice.ts';
+import { selectAuthIsAuth, selectAuthProfile } from '@/store/selectors.ts';
+import { authActions, getProfile, logoutUser } from '@/store/slices/authSlice.ts';
 import { useAppDispatch, useAppSelector } from '@/store/store.ts';
 
 const { Text } = Typography;
@@ -13,6 +14,13 @@ export const ProfilePage = () => {
   const navigate = useNavigate();
 
   const { data } = useAppSelector(selectAuthProfile);
+  const isAuth = useAppSelector(selectAuthIsAuth);
+
+  useEffect(() => {
+    if (isAuth) {
+      dispatch(getProfile());
+    }
+  }, [isAuth]);
 
   const onLogout = async () => {
     await dispatch(logoutUser()).unwrap();
