+++ src/store/utils.ts
@@ -28,7 +28,7 @@ export const getAsyncRequestData = <T>(
   stateParam: IAsyncParticle<T>
 ): {
   data: T | null;
-  error: string | null;
+  error: any;
   status: IAsyncDataStatus;
 } => ({
   data: stateParam?.data,
